+++ src/main/java/com/example/demo/dao/TArticleDao.java
@@ -22,7 +22,7 @@ public interface TArticleDao {
      * @param id 主键
      * @return 实例对象
      */
-    TArticle queryById(Integer id);
+    TArticle queryById(Long id);
 
     /**
      * 查询指定行数据
