+++ src/main/java/com/example/demo/dao/TCommentDao.java
@@ -21,7 +21,7 @@ public interface TCommentDao {
      * @param id 主键
      * @return 实例对象
      */
-    TComment queryById(Integer id);
+    TComment queryById(Long id);
 
     /**
      * 查询指定行数据
@@ -46,7 +46,7 @@ public interface TCommentDao {
      * @param articleId
      * @return
      */
-    List<TComment> queryAllByArticleId(Integer articleId);
+    List<TComment> queryAllByArticleId(Long articleId);
 
     /**
      * 新增数据
