+++ src/main/java/com/example/demo/dao/TImgDao.java
@@ -20,7 +20,7 @@ public interface TImgDao {
      * @param id 主键
      * @return 实例对象
      */
-    TImg queryById(Integer id);
+    TImg queryById(Long id);
 
     /**
      * 查询指定行数据
