+++ src/main/java/com/example/demo/dao/TSubscriptionDao.java
@@ -20,7 +20,7 @@ public interface TSubscriptionDao {
      * @param id 主键
      * @return 实例对象
      */
-    TSubscription queryById(Integer id);
+    TSubscription queryById(Long id);
 
     /**
      * 查询指定行数据
